+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 	U, umax = 55 (was 255)
 	V, vmax = 78 (was 255)
 - Added a documentation based on the Filter Factory documentation.
+- Added a help button which opens the documentation on the web. (*)
 
 (*) This bug/solution was tested on Windows but needs to be verified and/or implemented on Mac.